+++ sdk/storage/azure-storage-blob/azure/storage/blob/_shared_access_signature.py
@@ -153,7 +153,8 @@ class BlobSharedAccessSignature(SharedAccessSignature):
         :param ContainerSasPermissions permission:
             The permissions associated with the shared access signature. The
             user is restricted to operations allowed by the permissions.
-            Permissions must be ordered read, write, delete, list.
+            Permissions must be ordered read, write, delete, delete version, permanent delete, list, tag, find,
+            set immutability policy.
             Required unless an id is given referencing a stored access policy
             which contains this field. This field must be omitted if it has been
             specified in an associated stored access policy.
@@ -406,7 +407,8 @@ def generate_container_sas(
     :param permission:
         The permissions associated with the shared access signature. The
         user is restricted to operations allowed by the permissions.
-        Permissions must be ordered read, write, delete, list.
+        Permissions must be ordered read, write, delete, delete version, permanent delete, list, tag, find,
+        set immutability policy.
         Required unless an id is given referencing a stored access policy
         which contains this field. This field must be omitted if it has been
         specified in an associated stored access policy.