+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Release History
 
-## 12.10.0b3 (Unreleased)
+## 12.10.0b3 (2022-02-08)
 
 This version and all future versions will require Python 3.6+. Python 2.7 is no longer supported.
 
 ### Features Added
+- Added support for service version 2021-04-10.
+- Added support for `find_blobs_by_tags()` on a container.
+- Added support for `Find (f)` container SAS permission.
 
 ### Bugs Fixed
 - Update `azure-core` dependency to avoid inconsistent dependencies from being installed.